+++ libsrc4/nc4hdf.c
@@ -21,6 +21,11 @@
 #include <H5DSpublic.h>
 #include <math.h>
 
+#ifdef HAVE_INTTYPES_H
+#define __STDC_FORMAT_MACROS
+#include <inttypes.h>
+#endif
+
 #ifdef USE_PARALLEL
 #include "netcdf_par.h"
 #endif
@@ -1768,17 +1773,9 @@ var_create_dataset(NC_GRP_INFO_T *grp, NC_VAR_INFO_T *var, nc_bool_t write_dimid
       /* If there are no unlimited dims, and no filters, and the user
        * has not specified chunksizes, use contiguous variable for
        * better performance. */
-
-      if(!var->shuffle && !var->deflate &&
-         !var->fletcher32 && (var->chunksizes == NULL || !var->chunksizes[0])) {
-#ifdef USE_HDF4
-         NC_HDF5_FILE_INFO_T *h5 = grp->nc4_info;
-         if(h5->hdf4 || !unlimdim)
-#else
-            if(!unlimdim)
-#endif
-               var->contiguous = NC_TRUE;
-      }
+      if (!var->shuffle && !var->deflate && !var->fletcher32 &&
+          (var->chunksizes == NULL || !var->chunksizes[0]) && !unlimdim)
+         var->contiguous = NC_TRUE;
 
       /* Gather current & maximum dimension sizes, along with chunk sizes */
       for (d = 0; d < var->ndims; d++)
@@ -4226,10 +4223,10 @@ reportobject(int log, hid_t id, unsigned int type)
    }
    if(log) {
 #ifdef LOGGING
-      LOG((0,"Type = %s(%8u) name='%s'",typename,id,name));
+      LOG((0,"Type = %s(%8" PRId64 ") name='%s'",typename,id,name));
 #endif
    } else {
-      fprintf(stderr,"Type = %s(%8u) name='%s'",typename,(unsigned int)id,name);
+      fprintf(stderr,"Type = %s(%8" PRId64 ") name='%s'",typename,id,name);
    }
 }
 
@@ -4253,10 +4250,10 @@ reportopenobjectsT(int log, hid_t fid, int ntypes, unsigned int* otypes)
 
    if(log) {
 #ifdef LOGGING
-      LOG((0,"\nReport: open objects on %d\n",fid));
+      LOG((0,"\nReport: open objects on %" PRId64 "\n",fid));
 #endif
    } else {
-      fprintf(stdout,"\nReport: open objects on %d\n",(int)fid);
+      fprintf(stdout,"\nReport: open objects on %" PRId64 "\n",fid);
    }
    maxobjs = H5Fget_obj_count(fid,H5F_OBJ_ALL);
    if(idlist != NULL) free(idlist);